backend/requirements.txt
asgiref==3.3.1
certifi==2020.6.20
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
olefile==0.46
Pillow @ file:///C:/ci/pillow_1603823068645/work
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
wincertstore==0.2
